pyproject.toml
[tool.poetry]
name = "jango-restframework"
version = "0.1.0"
description = ""
authors = ["Saad Ahmed <saadmakhdoom12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
pygments = "^2.18.0"
django = "^5.1.4"
httpie = "^3.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
